# openapi.yaml
openapi: 3.1.0
info:
  title: Hooter Backend API
  version: '1.0.0'
  description: Internal API for Hooter backend
servers:
  - url: https://hooter_staging.h0oter.com
    description: Staging server

  - url: http://localhost:8888
    description: Local development server
...
paths:
...
  /catalog/images/{usku_id}:
    post:
      summary:  Images upload api for the product
      description: |
        Upload images for image type and image order for products.

        # Code Example
        ## React
        
        **Defining the images form storage**
        ```
        const [imagesData, setImagesData] = useState({});
        ```
        Here the images is in key: value pairs of `image-type`: {"image": `imageFile`, "order": `image-order`}
        > 0 <= `image-order` <= 20
      

        **Set image and metadata example**
        ```
        function storeImageData(e, type, order){
          const imageFile = e.target.files[0];

          setImagesData((prev)=> ({
            ...prev,
            [type]: {"order": order, "image": imageFile}
          }));
        }
        ```


        **Submit form function**
        ```
        async function submitForm(){
          const form = new FormData();
          const metadata = {};

          cosnt images = Object.entries(imagesData);
          images.forEach(([ type, { image, order } ])=>{
            form.append(type, image);   // <--- appending images
            metadata[type] = order      // <-- creating metadata
          });

          form.append("meta", JSON.stringify(metadata));  // <--- appending meta

          await uploadForm(form);
        }


        async function uploadForm(form){
          const response = await fetch(`api/route`, {
              method: 'POST',
              credentials: "included",
              body: form
            })
          
          const data = await response.json();
        }
        ```

        

        **Multipart form example**
        ```
        <div id="ImageForm">
          <input type=”file” name=front accept=”image/png,image/webp,image/jpeg,image/jpg”  onChange={(e)=> {storeImageData(e, imageType, imageOrder)}}/>
          <input type=”file” name=zoomed accept=”image/png,image/webp,image/jpeg,image/jpg”  onChange={(e)=> {storeImageData(e, imageType, imageOrder)}}/>

          <button type="submit" onClick=(submitForm)>Upload image</button>
        </div>
        ```

        > **Warning**
        > The form is supposed to be dynamically generated using the `image-attributes` from the category or custom images 
          and the image-type and image-order are supposed to be acquired from the user input if it's custom
      tags:
        - Catalog
      security:
        - cookieAuth: []
      parameters:
        - name: usku_id
          in: path
          required: true
          description: Universally unique Stock Keeping Unit ID for the products
          schema:
            type: string
        - $ref: '#/components/parameters/csrf'
      requestBody:
        description:  Images and their metadata, containing the image type and image order. The form payload will
          be according to the image attributes and any custom attributes for image that user has selected
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                meta:
                  type: string
                  description:  meta contains the metadata of the images according to their image types in 
                    the serielized JSON formate mapping `image_type` to `image_order` formate
                  example:  '{"front":3,"zoomed":6}'
                front:
                  type: string
                  format: binary
                zoomed:
                  type: string
                  format: binary
      responses:
        "200":
          description:  Successfully scheduled the image upload to run on the background
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example:  successful
                  message:
                    type: string
                    example:  Image upload has started
                  event_url:
                    type: string
                    example: /catalog/images/events/04c6521a4c4348a4ba12f8315c93e9c8
        "413":
          description:  Denied request for the image upload
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      status:
                        type: string
                      message:
                        type: string
                      allowed_size:
                        type: string

                  - type: object
                    properties:
                      status:
                        type: string
                      message:
                        type: string
                      image_type:
                        type: string 
              examples:
                size_not_allowed:
                  summary:  Image too large
                  value:
                    status: denied
                    message:  Image too large
                    allowed_size: 10MB
                Wrong_image_type:
                  summary:  Invalid file type
                  value:
                    status: failed
                    message:  File type should be an image
                    image_type: front
                Wrong_image_order:
                  summary:  Invalid file order
                  value:
                    status: failed
                    message:  Image order can not be equal to or less then 0
        "500":
          description:  Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: failed
                  message:
                    type: string
                    example:  Internal server error
        "default":
          $ref: '#/components/responses/default'
    get:
      summary:  Get product images urls
      description:  Get product images urls of the USKU ID, for original, high_resol, low_resol and webp card images.
      tags:
        - Catalog
      security:
        - cookieAuth: []
      parameters:
        - name: usku_id
          in: path
          required: true
          description: Universally unique Stock Keeping Unit ID for the products
          schema:
            type: string
        - name: image-type
          in: query
          required: false
          description:  Unique ID for category from shopify category taxnomomy
          schema:
            type: string
            example: front
      responses:
        "200":
          description:  Successfully received the image urls based upon the whether the image-type provided or not
          content:
            application/json:
              schema:
                oneOf:
                  - type: object
                    properties:
                      front:
                        type: object
                        properties:
                          image_order:
                            type: string
                          image_urls:
                            type: object
                            properties:
                              high_resol_webp:
                                type: string
                              low_resol_webp:
                                type: string
                              original:
                                type: string
                              webp_card:
                                type: string
                      zoomed:
                        type: object
                        properties:
                          image_order:
                            type: string
                          image_urls:
                            type: object
                            properties:
                              high_resol_webp:
                                type: string
                              low_resol_webp:
                                type: string
                              original:
                                type: string
                              webp_card:
                                type: string

                  - type: object
                    properties: 
                      image_order:
                        type: string
                      image_urls:
                        type: object
                        properties:
                          high_resol_webp:
                            type: string
                          low_resol_webp:
                            type: string
                          original:
                            type: string
                          webp_card:
                            type: string
              examples:
                image_type:
                  summary:  image-type urls
                  description:  Accessing the image-type specific keys
                  value:
                    image_order:  "3"
                    image_urls:
                      high_resol_webp: catalog/images/high_resol_webp/85448dd705574e3aafe07d10c58cca86/front.webp
                      low_resol_webp: catalog/images/low_resol_webp/85448dd705574e3aafe07d10c58cca86/front.webp
                      original: catalog/images/original_image/85448dd705574e3aafe07d10c58cca86/front..jpg
                      webp_card: catalog/images/webp_card/85448dd705574e3aafe07d10c58cca86/front.webp
                not_image_type:
                  summary:  All image urls
                  description:  Accessing all the image urls of the product
                  value:
                    front:
                      image_order: "3"
                      image_urls:
                        high_resol_webp: catalog/images/high_resol_webp/85448dd705574e3aafe07d10c58cca86/front.webp
                        low_resol_webp: catalog/images/low_resol_webp/85448dd705574e3aafe07d10c58cca86/front.webp
                        original: catalog/images/original_image/85448dd705574e3aafe07d10c58cca86/front..jpg
                        webp_card: catalog/images/webp_card/85448dd705574e3aafe07d10c58cca86/front.webp
                    zoomed:
                      image_order: "6"
                      image_urls:
                        high_resol_webp: catalog/images/high_resol_webp/85448dd705574e3aafe07d10c58cca86/zoomed.webp
                        low_resol_webp: catalog/images/low_resol_webp/85448dd705574e3aafe07d10c58cca86/zoomed.webp
                        original: catalog/images/original_image/85448dd705574e3aafe07d10c58cca86/zoomed..jpg
                        webp_card: catalog/images/webp_card/85448dd705574e3aafe07d10c58cca86/zoomed.webp
        "409":
          description:  Could not find image urls
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/default'
              examples:
                image_type:
                  summary:  image-type urls
                  description:  No image found while accessing the image-type specific keys
                  value:
                    status: failed
                    message:  Invalid image type or Image does not exists
                not_image_type:
                  summary:  All image urls
                  description:  No image found while accessing all the image urls of the product
                  value:
                    status: failed
                    message:  Image does not exists
        "default":
          $ref: "#/components/responses/default"
    delete:
      summary: Delete product images
      description: Delete product images based upon the `image-type` or all the images of the product
      security:
        - cookieAuth: []
      tags:
        - Catalog
      parameters:
        - name: usku_id
          in: path
          required: true
          description: Universally unique Stock Keeping Unit ID for the products
          schema:
            type: string
        - name: image-type
          in: query
          required: false
          description:  Unique ID for category from shopify category taxnomomy
          schema:
            type: string
            example: front
        - $ref: '#/components/parameters/csrf'
      responses:
        "200":
          description:  Successful request
          content:
            application/json:
              schema:
                $ref: '#/components/responses/default'
              example:
                status: successful
                message:  Successfully deleted the images
        "default":
          $ref: "#/components/responses/default"
    put:
      summary:  Update the product images
      description:  Update the images, including all the image quality variants, of the product
      security:
        - cookieAuth: []
      parameters:
        - in: path
          name: usku_id
          schema:
            $ref: "#/components/schemas/usku_id"
          required: true
          description: Universally unique SKU ID
        - $ref: '#/components/parameters/csrf'
      tags:
        - Catalog
      requestBody: 
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                front:
                  type: string
                  format: binary
      responses:
        "200":
          description:  Returns urls for the Server Side Event to track the upload status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                  event_url: 
                    type: string
              example:
                status: successful
                message: Image upload has started
                event_url: /catalog/images/events/rsdlwuf8cJS9LwOv04A__8CPKZMc0XIkK-vSQlBj6KA
        "default":
          $ref: "#/components/responses/default"
...
  /catalog/images/{key}:         
    get:
      summary:  Download product image
      description:  Download the product image url
      tags: 
        - Catalog
      parameters:
        - name: key
          in: path
          required: true
          description:  S3 object storage service key
          schema:
            type: string
      responses:
        "200":
          description:  Successful image download
          content:
            image/*:
              schema:
                type: string
                format: binary
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/default'
              example:
                status: failed
                message:  Image not found
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  schemas:
    usku_id:
      type: string
      description:  Univerally traces the stock accross all the brands
...
    default:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
...
  responses:
...
    "default":
      description: Default response
      content:
        application/json:
          schema:
            type: object
            required:
              - status
              - message
            properties:
              status:
                type: string
              message:
                type: string
  parameters:
...
    csrf:
        name: X-CSRF-Token
        in: header
        required: true
        description: CSRF acesss token acquired using /security/csrf API
        schema:
          type: string